Former-commit-id: 6bbf196a9568bdb38fa80c5ae23d7f7fbb404378

+++ unisim_simulators/cxx/vle4fuzr/emu.hh
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2019,
+ *  Copyright (c) 2019-2020,
  *  Commissariat a l'Energie Atomique (CEA)
  *  All rights reserved.
  *
@@ -41,7 +41,7 @@ struct Processor
     {
       using is_transparent = void;
       bool operator() (Page const& a, Page const& b) const { return a > b; }
-      bool operator() (Page const& a, unsigned b) const { return a.base > b; }
+      bool operator() (Page const& a, uint64_t b) const { return a.base > b; }
     };
     Page( Page&& page )
       : base(page.base), last(page.last), data(page.data)
